Extract search bar
