+++ lib/widgets/details_list_view.dart
