+++ lib/widgets/details_page_decoration.dart
