+++ lib/widgets/process_indicator.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class ProcessIndicator extends StatefulWidget {
+  const ProcessIndicator({super.key});
+
+  @override
+  State<ProcessIndicator> createState() => _ProcessIndicatorState();
+}
+
+class _ProcessIndicatorState extends State<ProcessIndicator> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 750,
+      decoration: const BoxDecoration(),
+      child: const Center(
+        child: CircularProgressIndicator(),
+      ),
+    );
+  }
+}
